chore: [] rq serverside rendering
SSR'ing:
The 404, index and slug page now prefetch all their RQ queries. There's a bunch of default queries that are prefetched, and then there's nested components that we have to fetch separately. There's login in the slug page that pulls components from the pages, and then prefetches what's inside them

Codegen:
The codegen has been tweaked to pull variables from a file that's injected in each RQ hook. It no longer relies on a custom fetcher, but a default fetches with the endpoint and params interpolated

Context & lib files:
The fetcherUrl and fetcherConfig have been pulled from the contentful-context, they now live in a new hook called `useExternalSpaceAndPreview`. It contains all the logic to fetch custom space and preview data in the client

prefetch-mappings.ts and prefetch-promise-array.ts have been added as utility files for the ssr prefetching. The map, maps Contentful models to codegenned queries.

Link component:
Tweaked the Link component so it now doesnt rely on an `as` prop to propagate search params to the URL. This helpes preserve the url params beyond navigation

fetcherConfig:
Added as single source of truth for RQ endpoint and fetch params. It exports a singleton that's mutable, which gets updated in for example the `useExternalSpaceAndPreview`.

fetcher.ts & fetcherParams.ts
The logic for these two files has been moved to `useExternalSpaceAndPreview`, it updates the fetchConfig from the abovemenetioned paragraph